updated hyperpixel drivers
changed hyperpixel driver to take a reference to peripherals instead of
stealing them itself to prevent issues in the event the peripherals are
re-used.

`hyperinit` was also changed to take an owned reference to self so it's
dropped once initialization is complete. this probably isn't necessary
but there's no reason to keep it around once the initialization is done.

if we do want to do more with it in the future, we can just change it
back to `&mut self`